Reformat report structure.
This makes reports a dict of lists of dicts, and updates the print
output function, as well as adding a plist output function and an
argument for switching between the two.

Not all reports are updated yet. Further, to support cruftiness, etc,
the reports are going to get refactored again.