Pylint, implement more accurate unused tests. Remove dead code.
Spruce now checks for name-version usage in manifests for determining
whether an item is used.

Spruce also now considers the update_for and requires values of pkginfos
for items that may not be in manifests, but are thus still used. It
does not yet check _those_ for the name-version construction.