Add proof of concept docs verb.
Implementation at this point builds markdown output of each unique
product/item in the repo and outputs (to stdout) a table of name,
display name, all versions, and any notes that may be present.

argparse handling of output directory and html output flag are in place.

markdown package is imported with potential to safely fail.