Implement more accurate out-of-date checking code.
This version ensures that, for each OS version supported, there are X
versions saved for each specified item in a manifest, its required
items, and any updates for that item. Further refactoring will
potentially allow for further filtering.

The NoUsageReport and OutOfDate report were edited to use the new code.
Final location of this class and cleanup of now-dead code is ongoing.